backend/requirements.txt
pillow==9.5.0
ultralytics==8.0.145
fastapi==0.100.1
uvicorn==0.23.2
python-multipart==0.0.6